+++ src/controller/controldevice/controller/mapping/sdl/SDLButtonToButtonMapping.cpp
@@ -31,7 +31,7 @@ uint8_t SDLButtonToButtonMapping::GetMappingType() {
 }
 
 std::string SDLButtonToButtonMapping::GetButtonMappingId() {
-    return StringHelper::Sprintf("P%d-B%d-LUSI%d-SDLB%d", mPortIndex, mBitmask,
+    return StringHelper::Sprintf("P%d-B%u-LUSI%d-SDLB%d", mPortIndex, static_cast<uint32_t>(mBitmask),
                                  ControllerInputMapping::mShipDeviceIndex, mControllerButton);
 }
